Move all JavaScript logic to dedicated files
