+++ src/index.js
@@ -1,9 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
-
+import App from './components/App';
+import './styles/index.scss';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-    <App/>
-);
+root.render(<App />);
